requirements.txt
Django==5.2.8
pillow==12.0.0
djangorestframework==3.16.1